only author can delete comment

+++ controllers/post.js
@@ -151,6 +151,8 @@ const postController = {
                 res.code = 404;
                 throw new Error("Posts not found")
             }
+
+            if(post.)
     
             await Post.findByIdAndDelete(id)
             res.status(200).json({code : 200, status : true, message : "Post deleted successfully"})
